Added Observer to the whole app
