+++ app_feup/lib/main.dart
@@ -16,18 +16,29 @@ import 'package:redux/redux.dart';
 import 'package:redux_thunk/redux_thunk.dart';
 import 'redux/reducers.dart';
 import 'controller/parsers/parser-exams.dart';
+import 'package:app_feup/controller/LifecycleEventHandler.dart';
 
 List<Exam> exams;
 
 void main() => runApp(new MyApp());
 
-class MyApp extends StatelessWidget {
+class MyApp extends StatefulWidget {
+  @override
+  State<StatefulWidget> createState() {
+    return MyAppState();
+  }
+
+}
+
+class MyAppState extends State<MyApp> {
 
   final Store<AppState> state = Store<AppState>(
     appReducers, /* Function defined in the reducers file */
     initialState: new AppState(null),
     middleware: [generalMiddleware]
   );
+
+  WidgetsBindingObserver lifeCycleEventHandler;
 
   @override
   Widget build(BuildContext context) {
@@ -51,4 +62,17 @@ class MyApp extends StatelessWidget {
         },
     )
   );}
+
+  @override
+  void initState() {
+    super.initState();
+    this.lifeCycleEventHandler = new LifecycleEventHandler(store: this.state);
+    WidgetsBinding.instance.addObserver(this.lifeCycleEventHandler);
+  }
+
+  @override
+  void dispose() {
+    WidgetsBinding.instance.removeObserver(this.lifeCycleEventHandler);
+    super.dispose();
+  }
 }
